+++ CHANGELOG.md
@@ -21,7 +21,7 @@ For more information on SemVer, please visit http://semver.org/.
 ---
 
 ## Version 1.7.3
-### 2014-11-??
+### 2014-11-28
 
 * Added 14 news font icons
 * Improved WPDK Preferences with `wpdk_preferences_reset_to_default_branch-{$branch}`.
@@ -31,7 +31,6 @@ For more information on SemVer, please visit http://semver.org/.
 * Fixed dynamic table css styles.
 * Fixed potential Javascript warning in localize script on jQuery
 * Minor stability fixes
-
 
 ## Version 1.7.2
 ### 2014-11-19